chore(): Organise components
