+++ components/InputChannel/InputChannel.vue
@@ -36,8 +36,8 @@ import Vue from 'vue'
 import _isEmpty from 'lodash/isEmpty'
 import _cloneDeep from 'lodash/cloneDeep'
 import _startsWith from 'lodash/startsWith'
-import DioButton from '../DioButton.vue'
-import InputText from './InputText.vue'
+import DioButton from '../DioButton/DioButton.vue'
+import InputText from '../InputText/InputText.vue'
 
 const DEFAULT_URLPREFIX = 'https://'
 const DEFAULT_EMAILPREFIX = 'mailto://'
