+++ nuxt.config.js
@@ -69,6 +69,9 @@ export default {
 
   // Build Configuration: https://go.nuxtjs.dev/config-build
   build: {
+    babel: {
+      plugins: []
+    },
     postcss: {
       plugins: {
         // "postcss-css-variables": {},
