+++ package.json
@@ -23,7 +23,7 @@
     "lodash": "^4.17.21",
     "markdown-it-attrs": "^4.0.0",
     "markdown-it-div": "^1.1.0",
-    "nuxt": "^2.15.3",
+    "nuxt": "^2.15.7",
     "nuxt-i18n": "^6.27.0",
     "vue-country-flag": "^2.1.1"
   },
