+++ pages/policymaker/introduction.vue
@@ -14,10 +14,10 @@
 
 <script lang="ts">
 import Vue from 'vue'
-import ProgressSteps from '@/components/ProgressSteps.vue'
-import ProgressStep from '@/components/ProgressStep.vue'
-import PageTitle from '~/components/PageTitle.vue'
-import DioLink from '~/components/DioLink.vue'
+import ProgressSteps from '~/components/DioProgressStep/ProgressSteps.vue'
+import ProgressStep from '~/components/DioProgressStep/ProgressStep.vue'
+import PageTitle from '~/components/PageTitle/PageTitle.vue'
+import DioLink from '~/components/DioLink/DioLink.vue'
 import { store } from '~/store'
 
 export default Vue.extend({
